[svn-r1108] Changes since 19990225 ----------------------
./INSTALL
	Added warnings and a disclaimer about GNU, DEC, Irix64, and NT
	compilers that generate incorrect machine code.

./configure.in
./src/H5private.h
	Detects and includes <sys/param.h> which is needed on FreeBSD
	before <sys/proc.h> even though we only really using anything
	from <sys/proc.h> on the DEC Alpha.

./config/irix64
	Turned off warnings for duplicate definitions from the linker
	because -lnsl on irix has the same stuff in it as -lc.

./config/irix6.x
	Split up the CC and CFLAGS settings like with irix64 so that
	compilers besides `-n32' can be used.

./bin/snapshot
	The snapshots are tagged with names like hdf5-1_1_52 which is
	similar to the way the releases are tagged (cvs doesn't allow
	dots in tags).

./test/dtypes.c
./src/H5private.h
./src/H5Tconv.c
	Fixed some alignment violations on the DEC when using high
	optimization levels. The DEC incorrectly optimizes certain
	memcpy() and memmove() calls when the source argument is not
	word aligned if the call looks like it's copying an atomic
	data type.

./test/hyperslab
	Worked around code generation bugs in the Irix64 Mongoose 7.00
	compiler by casting some `unsigned long' values to `unsigned'
	in an expression.

./src/H5Ocomp.c
	Fixed a place where %d was used to print a size_t.

+++ src/H5Ocomp.c
@@ -449,7 +449,7 @@ H5O_pline_debug (H5F_t __unused__ *f, const void *mesg, FILE *stream,
 		(unsigned long)(pline->filter[i].cd_nelmts));
 	for (j=0; j<pline->filter[i].cd_nelmts; j++) {
 	    char	field_name[32];
-	    sprintf(field_name, "CD value %d", j);
+	    sprintf(field_name, "CD value %lu", (unsigned long)j);
 	    fprintf(stream, "%*s%-*s %lu\n", indent+6, "", MAX(0, fwidth-6),
 		    field_name,
 		    (unsigned long)(pline->filter[i].cd_values[j]));
